wip - moving the hop records into their own component
