+++ src/constants.js
@@ -1,0 +1,4 @@
+const IBU_INTERMEDIATE_GRAVITY = "intermediate-gravity";
+const IBU_FINAL_GRAVITY = "final-gravity";
+
+export { IBU_INTERMEDIATE_GRAVITY, IBU_FINAL_GRAVITY };
